Refactoring components directory
