+++ src/components/User/Login.vue
@@ -37,7 +37,7 @@
 </template>
 
 <script>
-import { baseURI } from './constants';
+import { baseURI } from '../constants';
 import VueRecaptcha from 'vue-recaptcha';
 
 export default {
